Refactor: 로그인 공통 로직 외의 기능 추가
공통로직 파일 외의 로그인 페이지에만 사용하는 함수를 로그인페이지에 정의해주었습니다.
