+++ signup.html
@@ -35,7 +35,7 @@
         </div>
       </div>  
     </div>  
-    <script src="js/login.js" type="module"></script>
+    <script src="js/signin_up/signup.js" type="module"></script>
   </body>
     
 </html>
